Fixing Modules & Adding in Group Concept
It turned out that the module changes weren't actually reloading. Once I
loaded them, there were bugs. Now the migration to a single namespace is
complete.

Part of the problem with the various faces was related to the way I was
drawing. Originally I was just putting everything on the model. Now I'm
adding everything to the group that contains the original face. This
gets SketchUp to help me keep track of the new faces I'm creating.